+++ packages/devtools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 1.0.0-rc.1
+
+### Patch Changes
+
+- 71f33a761: Final release of all RC ima@19 packages
+
 ## 1.0.0-rc.0
 
 ### Major Changes